Master: fixes insert movies, enhances initiate script to also retrieve csv, enhances readme

+++ mongo/mongo_insert_movies.py
@@ -1,4 +1,5 @@
 import json
+
 from mongo_connection import collection
 
 # Read Json
@@ -7,7 +8,21 @@ with open('ratings.json', 'r') as f:
 ratings = json.loads(raw_data)
 
 # Insert Data
-# TODO replace this to only insert if not exist
-collection.drop()
-result = collection.insert_many(ratings)
 
+if collection.find().count():
+    for rating in ratings:
+        movie = collection.find_one({'movie.id': rating['movie']['id']})
+        if movie:
+            # if the movie is already in the collection, update the score
+            collection.update_one({'_id': movie['_id']},
+                                  {'$set': {'score': movie['score']},
+                                   "$currentDate": {"recordLastModified": True}
+                                   })
+        else:
+            collection.insert_one(rating)
+
+
+else:
+    # If there are no records, drop and bulk insert is faster
+    collection.drop()
+    result = collection.insert_many(ratings)
